No longer depends entirely on barcharts
