+++ src/client/app/actions/graph.js
@@ -5,8 +5,7 @@
  */
 
 import { fetchNeededLineReadings } from './lineReadings';
-import { fetchNeededBarReadings } from './barReadings';
-import { fetchNeededCompareReadings } from './compareReadings';
+import { fetchNeededBarReadings, fetchNeededCompareReadings } from './barReadings';
 
 export const UPDATE_SELECTED_METERS = 'UPDATE_SELECTED_METERS';
 export const UPDATE_BAR_DURATION = 'UPDATE_BAR_DURATION';
@@ -49,7 +48,7 @@ export function changeSelectedMeters(meterIDs) {
 		dispatch(dispatch2 => {
 			dispatch2(fetchNeededLineReadings(getState().graph.timeInterval));
 			dispatch2(fetchNeededBarReadings(getState().graph.timeInterval));
-			dispatch2(fetchNeededCompareReadings(getState().graph.timeInterval));
+			dispatch2(fetchNeededCompareReadings(getState().graph.compareTimeInterval));
 		});
 		return Promise.resolve();
 	};
